chore(captp): rename lib -> src
