+++ packages/captp/jsconfig.json
@@ -14,5 +14,5 @@
     "strictNullChecks": true,
     "moduleResolution": "node",
   },
-  "include": ["lib/**/*.js", "exported.js"],
+  "include": ["src/**/*.js", "exported.js"],
 }
