+++ packages/captp/test/test-trap.js
@@ -3,7 +3,7 @@ import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava';
 
 import { Worker } from 'worker_threads';
 
-import { E, makeLoopback } from '../lib/captp';
+import { E, makeLoopback } from '../src/captp';
 import {
   createHostBootstrap,
   makeGuest,
